% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projectP2L}
\alias{projectP2L}
\title{Project a point (x, y) onto a line (a, b)}
\usage{
projectP2L(Pxy, Lab, add = FALSE, d0 = FALSE, color = "blue", addLine = FALSE)
}
\arguments{
\item{Pxy}{A vector c(x,y) with coordinates of a point.}

\item{Lab}{A vector c(a,b) with intercept (a) and slope (b) of a line.}

\item{add}{Logical: Should the projection be drawn (default = FALSE).}

\item{d0}{Logical: Should the distance line from a to
projection be drawn (default = FALSE).}

\item{color}{Character - a color name}

\item{addLine}{Logical: Should the line be drawn.}
}
\value{
A list with the coordinates of the projected point, and
the distance vector of the projected point.
}
\description{
On an existing plot including a line with a = intercept and b = slope,
a point with coordinates (x, y) is projected. Optionally, the line
as well as the projection arrow and the distance from intercept to
the projected point are drawn.
}
\examples{
\dontrun{
plot(-2:2, -2:2, type = "n", asp = 1)
abline(a = 0, b = 1)
aPoint <- matrix(c(-1.5, 1), nrow = 1, byrow = TRUE)
points(aPoint, pch = 19)
projectP2L(Pxy = aPoint, Lab = c(0, 1), add = TRUE)
}

}
