% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gutt91}
\alias{gutt91}
\title{Guttman 1991 Intellligence Data}
\format{
A list with two elements: gutt91_cor, gutt91_var
\describe{
\item{gutt91_cor}{A correlation matrix of 12 intelligence items}
\item{gutt91_var}{A data frame with 12 observations and 3 variables:
items = short name of the item, Material = Facet 1 (Oral, Manual, Paper),
Modality = Facet 2 (verbal, numerical, figural)}
}
}
\source{
smacof package, data(Guttman1991)
}
\usage{
gutt91
}
\description{
Data are based on the smacof package. It contains the correlations
among 12 intelligence items, and the assignment to two facets:
Material and Modality.
}
\examples{
data(gutt91)
Kor <- gutt91$gutt91_cor
Facets <- gutt91$gutt91_var
# convert correlations to distances
Kor_D <- smacof::sim2diss(Kor, method = "corr", to.dist = TRUE)
# MDS
gutt91_mds <- smacof::mds(Kor_D, type = "ordinal")
plot(gutt91_mds, main = "Guttman 1991 Intelligence")
# Angular partition of Modality
angularPartition(crd = gutt91_mds$conf, group = Facets$Modality)

}
\references{
Guttman, L. & Levy, S. (1991). Two structural laws for intelligence tests. Intelligence, 15, 79-103.
}
\keyword{datasets}
